requirements.txt
aes==1.2.0
asgiref==3.7.2
certifi==2023.5.7
charset-normalizer==3.1.0
crypto==1.4.1
Crypto.Cipher==1
Django==3.0
django-admin-thumbnails==0.2.8
django-ckeditor==6.5.1
django-crispy-forms==2.0
django-crypto==0.2.0
django-des==2.4.1
django-js-asset==2.0.0
django-mathfilters==1.0.0
django-mptt==0.14.0
django-solo==1.1.3
idna==3.4
Naked==0.1.32
numpy==1.24.3
pandas==2.0.2
Pillow==9.5.0
pycrypto==2.6.1
pycryptodome==3.18.0
pycryptodomex==3.18.0
python-dateutil==2.8.2
python2==1.2
pytz==2023.3
PyYAML==6.0
requests==2.31.0
shellescape==3.8.1
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.6.3
tzdata==2023.3
urllib3==2.0.3
